Move text-related elements to the 'text' element group folder
