+++ angular/components/elements/element-group-text/subtitle/subtitle.component.scss
@@ -1,0 +1,12 @@
+// subtitle.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+subtitle {
+    @include flex(center, center, row);
+    color: $absolute-black;
+    font-style: italic;
+    text-align: center;
+}
