+++ angular/components/elements/element-group-text/subtitle/subtitle.component.ts
@@ -1,6 +1,6 @@
 // subtitle.component.ts 
 import {Component, Input} from "@angular/core";
-import {TagType} from "../../../models/html/TagType";
+import {TagType} from "../../../../models/html/TagType";
 
 @Component({
     selector: 'subtitle',
