--- angular/components/elements/subtitle/subtitle.component.scss
@@ -1,12 +1,0 @@
-// subtitle.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-subtitle {
-    @include flex(center, center, row);
-    color: $absolute-black;
-    font-style: italic;
-    text-align: center;
-}
